backend/requirements.txt
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0